1.save and read draw data, update draw data 2.make all kinds of id as Long

+++ app/src/main/java/com/king/app/plate/model/db/dao/MatchDao.kt
@@ -16,5 +16,5 @@ interface MatchDao:BaseDao<Match> {
     fun getMatches(): List<Match>
 
     @Query("select * from `match` where id=:id")
-    fun getMatchById(id: Int): Match
+    fun getMatchById(id: Long): Match
 }
